requirements.txt
fastapi
uvicorn
numpy
librosa
soundfile
tensorflow-cpu
python-multipart
h5py